python/requirements.txt
termcolor
tqdm
mysql-connector-python
pymongo
